+++ transports/websocket/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.46.0
 
+- Add `tls::Config::new_with_server_cert_resolver` to resolve server certificates dynamically.
+  See [PR XXXX](https://github.com/libp2p/rust-libp2p/pull/XXXX).
 - Raise MSRV to 1.88.0.
   See [PR 6273](https://github.com/libp2p/rust-libp2p/pull/6273).
 - feat(websocket): support `/tls/sni/<hostname>/ws` multiaddrs in the WebSocket transport